CreateTable: add SMALLINT | Remove ColumnSpecification (#200)
PR #177 added additional column types, including `SMALLINT`. However, SMALLINT was not added as a token for the lexical analysis (flex scanner). This PR adds `SMALLINT` as a token and extends a test case so that all column types are tested with a `CREATE TABLE` SQL statement.

Furthermore, this PR eliminates the `ColumnSpecification` struct, which was used before to store precision and scale information for the DECIMAL or TIME column types. Similar to the `ColumnType::length`, `precision` and `scale` are now member variables of `ColumnType`.