INTERVAL support, date pseudo literals TIMESTAMP and TIME literals.
The interval was not allowed to be a column data type.
With this patch we add this support.

We add support for the pseudo literals
CURRENT_TIME, CURRENT_TIMESTAMP, CURRENT_DATE.

Also, we add support for TIMESTAMP and TIME literals.
The user can now explicitly set the type of the string value it follows
with TIMESTAMP 'value' and TIME 'value.

DATE was already supported.