CSV import options for DELIMITER, NULL, and QUOTE (#256)
This branch implements more csv parsing options. The goal is that the
following sql string is correctly parsed:
```
COPY "{}" FROM '{}' WITH (FORMAT CSV, DELIMITER '|', NULL '', QUOTE '"');
```

where delimiter is the delimiter of the csv file, null is the value of
null values in the csv and quote is the string used to quote values.